feat(task): Extract todoist updates
