+++ src/components/tasks/TodoistTasks.vue
@@ -1,0 +1,44 @@
+<script>
+import { watch } from 'vue';
+import { useStore } from 'vuex';
+import useSWRV from 'swrv';
+import axios from 'axios';
+
+export default {
+  setup() {
+    const store = useStore();
+    const allTaskURL = 'https://api.todoist.com/rest/v1/tasks?filter=today';
+    const allProjectsURL = 'https://api.todoist.com/rest/v1/projects';
+    const todoistKey = process.env.VUE_APP_TODOISTKEY;
+
+    function fetcher(url) {
+      return axios
+        .get(url, { data: {}, headers: { Authorization: `Bearer ${todoistKey}` } })
+        .then((response) => response.data)
+        .catch((error) => {
+          console.error(error);
+        });
+    }
+
+    const addTasks = (newTasks) =>
+      store.dispatch('addTasks', { source: 'Todoist', tasks: newTasks });
+
+    // Get todoist data
+    const { data: todoistTasks, error: taskError } = useSWRV(allTaskURL, fetcher);
+    // Get todoist projects
+    const { data: todoistProjects, error: projectsError } = useSWRV(allProjectsURL, fetcher);
+
+    watch(todoistTasks, () => {
+      addTasks(todoistTasks.value);
+    });
+
+    return {
+      allTaskURL,
+      todoistKey,
+      taskError,
+      todoistProjects,
+      projectsError,
+    };
+  },
+};
+</script>
